% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR.R
\name{irr}
\alias{irr}
\title{Compute the Internal Rate of Return (IRR) of an annual coupon payment bond}
\usage{
irr(nominal = nominal, Mat_Date = Mat_Date, coupon = coupon,
  Buy_Date = Buy_Date, price = price)
}
\arguments{
\item{nominal}{nominal of the bond}

\item{Mat_Date}{maturity date}

\item{coupon}{coupon in percentage of nominal}

\item{Buy_Date}{Date of buy, with %Y-%m-%d format}

\item{price}{bond price at the buy date}
}
\value{
IRR of the bond
}
\description{
Compute the Internal Rate of Return (IRR) of an annual coupon payment bond
}
